requirements.txt
bson==0.5.8
dataclasses==0.6
Django==2.2.11
djangorestframework==3.11.0
djongo==1.3.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
